3e config des bras

+++ robot-system-lib-parent/robot-system-lib-core/src/main/java/org/arig/robot/model/bras/AnglesBras.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class AnglesBras {
-    public final int a1;
-    public final int a2;
-    public final int a3;
+    public final double a1;
+    public final double a2;
+    public final double a3;
     public boolean a1Error;
     public boolean a2Error;
     public boolean a3Error;
